[BIA-560] Postgres: Deleting resource is not deleting references and alias
Update delete and update modules.
Update test to validate if aliases and references are deleted when a document is deleted.
Update update module to use an old meadowlarkId to filter aliases and references, because, if it change, this query will return empty.
Add validation to use enable or disable allow Identity Updates